+++ gcovr/__main__.py
@@ -246,24 +246,7 @@ def main(args=None):
         sys.exit(1)
 
     if options.objdir is not None:
-        if not options.objdir:
-            logger.error(
-                "empty --object-directory option.\n"
-                "\tThis option specifies the path to the object file "
-                "directory of your project.\n"
-                "\tThis option cannot be an empty string."
-            )
-            sys.exit(1)
-        tmp = options.objdir.replace("/", os.sep).replace("\\", os.sep)
-        while os.sep + os.sep in tmp:
-            tmp = tmp.replace(os.sep + os.sep, os.sep)
-        if normpath(options.objdir) != tmp:
-            logger.warning(
-                "relative referencing in --object-directory.\n"
-                "\tthis could cause strange errors when gcovr attempts to\n"
-                "\tidentify the original gcc working directory."
-            )
-        if not os.path.exists(normpath(options.objdir)):
+        if not os.path.exists(options.objdir):
             logger.error(
                 "Bad --object-directory option.\n"
                 "\tThe specified directory does not exist."
@@ -271,14 +254,6 @@ def main(args=None):
             sys.exit(1)
 
     options.starting_dir = os.path.abspath(os.getcwd())
-    if not options.root:
-        logger.error(
-            "empty --root option.\n"
-            "\tRoot specifies the path to the root "
-            "directory of your project.\n"
-            "\tThis option cannot be an empty string."
-        )
-        sys.exit(1)
     options.root_dir = os.path.abspath(options.root)
 
     #