+++ gcovr/configuration.py
@@ -69,6 +69,22 @@ def check_input_file(value: str, basedir: str = None) -> str:
     return os.path.abspath(value)
 
 
+def relative_path(value: str, basedir: str = None) -> str:
+    r"""
+    Make a absolute path if value is a relative path.
+    """
+    if not value:
+        raise ArgumentTypeError("Should not be set to an empty string.") from None
+
+    if basedir is None:
+        basedir = os.getcwd()
+
+    if not os.path.isabs(value):
+        value = os.path.join(basedir, value)
+    value = os.path.normpath(value)
+    return os.path.relpath(value, os.getcwd())
+
+
 def timestamp(value: str) -> datetime.datetime:
     from .timestamps import parse_timestamp  # lazy import
 
@@ -478,6 +494,12 @@ def _get_value_from_config_entry(
         except (ValueError, ArgumentTypeError) as err:
             raise cfg_entry.error(str(err))
 
+    elif option.config == "add-tracefile":  # Special case for patterns
+        assert (
+            cfg_entry.filename is not None
+        ), "conversion function must derive base directory from filename"
+        basedir = os.path.dirname(cfg_entry.filename)
+        value = os.path.join(basedir, cfg_entry.value)
     else:
         value = cfg_entry.value
 
@@ -510,6 +532,9 @@ def _get_converter_function(
 
     if option_type is check_input_file:
         return lambda value: check_input_file(value, basedir)
+
+    if option_type is relative_path:
+        return lambda value: relative_path(value, basedir)
 
     if option_type is OutputOrDefault:
         return lambda value: OutputOrDefault(value, basedir)
@@ -626,6 +651,7 @@ GCOVR_CONFIG_OPTIONS = [
             "The --root is the default --filter."
         ),
         default=".",
+        type=relative_path,
     ),
     GcovrConfigOption(
         "add_tracefile",
@@ -654,6 +680,7 @@ GCOVR_CONFIG_OPTIONS = [
             "Search these directories for coverage files. "
             "Defaults to --root and --object-directory."
         ),
+        type=relative_path,
     ),
     GcovrConfigOption(
         "config",
@@ -663,6 +690,7 @@ GCOVR_CONFIG_OPTIONS = [
             "Load that configuration file. "
             "Defaults to gcovr.cfg in the --root directory."
         ),
+        type=relative_path,
     ),
     GcovrConfigOption(
         "respect_exclusion_markers",
@@ -1266,6 +1294,7 @@ GCOVR_CONFIG_OPTIONS = [
             "the path from gcc to the gcda file (i.e. gcc's '-o' option), "
             "or the path from the gcda file to gcc's working directory."
         ),
+        type=relative_path,
     ),
     GcovrConfigOption(
         "keep",